Inlining some code and templating least squares routines
