+++ Source/AmrMesh/CD_EBMultigridInterpolator.cpp
@@ -530,9 +530,9 @@ bool EBMultigridInterpolator::getStencil(VoFStencil&            a_stencilFine,
   const int numEquations = coarVofs.size() + fineVofs.size();
   const int numUnknowns  = LeastSquares::getTaylorExpansionSize(a_order);
 
-  timer.startEvent("lsq solve");
-  if(numEquations >= numUnknowns) { // We have enough equations to get a stencil.
 
+  if(numEquations >= numUnknowns) { // We have enough equations to get a stencil.
+    timer.startEvent("sort");
     // In many cases we will have WAY too many equations for the specified order. This is particularly true in 3D
     // because the number of coar vofs included in a radius r from the ghost vof can be (1 + 2*r)^3. So for r = 2
     // this = 125 cells, not counting the fine cells. Since singular value decomposition scales like O(n^3), the
@@ -581,11 +581,13 @@ bool EBMultigridInterpolator::getStencil(VoFStencil&            a_stencilFine,
     
     fineVofsTrimmedSize.resize(std::min(2*numUnknowns, curFineSize));
     coarVofsTrimmedSize.resize(std::min(2*numUnknowns, curCoarSize));
+    timer.stopEvent("sort");    
     
     // Build displacement vectors
     Vector<RealVect> fineDisplacements;
     Vector<RealVect> coarDisplacements;
 
+    timer.startEvent("dist");
     for (const auto& fineVof : fineVofs.stdVector()){
       fineDisplacements.push_back(LeastSquares::displacement(a_dataLocation, a_dataLocation, a_ghostVofFine, fineVof, a_ebisboxFine, a_dxFine));
     }
@@ -593,6 +595,7 @@ bool EBMultigridInterpolator::getStencil(VoFStencil&            a_stencilFine,
     for (const auto& coarVof : coarVofs.stdVector()){
       coarDisplacements.push_back(LeastSquares::displacement(a_dataLocation, a_dataLocation, a_ghostVofFine, coarVof, a_ebisboxFine, a_ebisboxCoar, a_dxFine, a_dxCoar));
     }
+    timer.stopEvent("dist");    
 
     // LeastSquares computes all unknown terms in a Taylor expansion up to specified order. We want the 0th order term, i.e. the interpolated value,
     // which in multi-index notation is the term (0,0), i.e. IntVect::Zero. The format of the two-level least squares routine is such that the
@@ -601,17 +604,20 @@ bool EBMultigridInterpolator::getStencil(VoFStencil&            a_stencilFine,
     IntVectSet derivs       = IntVectSet(interpStenIndex);
     IntVectSet knownTerms   = IntVectSet();
 
-    std::map<IntVect, std::pair<VoFStencil, VoFStencil> > stencils = LeastSquares::computeDualLevelStencils(derivs,
-													    knownTerms,
-													    fineVofs,
-													    coarVofs,
-													    fineDisplacements,
-													    coarDisplacements,
-													    a_weight,
-													    a_order);
+    timer.startEvent("stencil");
+    std::map<IntVect, std::pair<VoFStencil, VoFStencil> > stencils
+      = LeastSquares::computeDualLevelStencils<float>(derivs,
+						      knownTerms,
+						      fineVofs,
+						      coarVofs,
+						      fineDisplacements,
+						      coarDisplacements,
+						      a_weight,
+						      a_order);
 
     a_stencilFine = stencils.at(interpStenIndex).first;
     a_stencilCoar = stencils.at(interpStenIndex).second;
+    timer.stopEvent("stencil");    
 
     foundStencil = true;
   }
@@ -619,7 +625,6 @@ bool EBMultigridInterpolator::getStencil(VoFStencil&            a_stencilFine,
     std::cout << "/b/" << numEquations << "/" << numUnknowns << "/e/\n";    
     foundStencil = false;
   }
-  timer.stopEvent("lsq solve");
 
   timer.eventReport(true);
