+++ Source/Utilities/CD_TimerImplem.H
@@ -110,8 +110,8 @@ void Timer::printReportTail(const std::pair<Real, Real> a_elapsedTime) const noe
   std::stringstream tail;
   
   tail << "| ---------------------------------------------------------------------------------------------------------|" << "\n"
-       << "| Local elapsed time  = " << std::fixed << std::setprecision(3) << a_elapsedTime.first  << " seconds"         << "\n"    
-       << "| Global elapsed time = " << std::fixed << std::setprecision(3) << a_elapsedTime.second << " seconds"         << "\n"
+       << "| Local elapsed time  = " << std::fixed << std::setprecision(5) << a_elapsedTime.first  << " seconds"         << "\n"    
+       << "| Global elapsed time = " << std::fixed << std::setprecision(5) << a_elapsedTime.second << " seconds"         << "\n"
        << "| ---------------------------------------------------------------------------------------------------------|" << "\n";
 
   pout() << tail.str();
@@ -146,8 +146,8 @@ void Timer::eventReport(const bool a_localReportOnly) const noexcept{
       const Duration localDuration      = std::get<ElapsedTime>(e.second);
       const Real localDurationInSeconds = localDuration.count();
 
-      ssLocalDuration << std::fixed << std::setprecision(3) << localDurationInSeconds;
-      ssPercentage    << std::fixed << std::setprecision(3) << 100.*localDurationInSeconds/totalTimeLocal;
+      ssLocalDuration << std::fixed << std::setprecision(5) << localDurationInSeconds;
+      ssPercentage    << std::fixed << std::setprecision(5) << 100.*localDurationInSeconds/totalTimeLocal;
 
 #ifdef CH_MPI
       if(!a_localReportOnly){
@@ -199,12 +199,12 @@ void Timer::eventReport(const bool a_localReportOnly) const noexcept{
 	broadcast(minRank, srcRank);
 	broadcast(maxRank, srcRank);      
 
-	ssMinDuration << std::fixed << std::setprecision(3) << minTime;
-	ssMaxDuration << std::fixed << std::setprecision(3) << maxTime;
-	ssAvgDuration << std::fixed << std::setprecision(3) << average;
-	ssDevDuration << std::fixed << std::setprecision(3) << sigma;
-	ssMinRank     << std::fixed << std::setprecision(3) << minRank;
-	ssMaxRank     << std::fixed << std::setprecision(3) << maxRank;
+	ssMinDuration << std::fixed << std::setprecision(5) << minTime;
+	ssMaxDuration << std::fixed << std::setprecision(5) << maxTime;
+	ssAvgDuration << std::fixed << std::setprecision(5) << average;
+	ssDevDuration << std::fixed << std::setprecision(5) << sigma;
+	ssMinRank     << std::fixed << std::setprecision(5) << minRank;
+	ssMaxRank     << std::fixed << std::setprecision(5) << maxRank;
       }
       else{
 	ssMinDuration    << " N/A ";
